Restructure repo into directories as Lit-LLaMA
