+++ howto/finetune_adapter.md
@@ -24,7 +24,7 @@ or [prepare your own dataset](#tune-on-your-dataset).
 ## Running the finetuning
 
 ```bash
-python finetune_adapter.py --checkpoint_dir checkpoints/stabilityai/stablelm-base-alpha-3b
+python finetune/adapter.py --checkpoint_dir checkpoints/stabilityai/stablelm-base-alpha-3b
 ```
 
 The finetuning requires at least one GPU with ~12 GB memory.
@@ -40,7 +40,7 @@ micro_batch_size = 4
 This script will save checkpoints periodically to the `out_dir` directory. If you are finetuning different models or on your own dataset, you can specify an output directory with your preferred name:
 
 ```bash
-python finetune_adapter.py --out_dir out/adapter/my-model-finetuned
+python finetune/adapter.py --out_dir out/adapter/my-model-finetuned
 ```
 
 ## Test the model
@@ -48,7 +48,7 @@ python finetune_adapter.py --out_dir out/adapter/my-model-finetuned
 You can test the finetuned model with your own instructions by running:
 
 ```bash
-python generate_adapter.py \
+python generate/adapter.py \
     --prompt "Recommend a movie to watch on the weekend." \
     --checkpoint_dir checkpoints/stabilityai/stablelm-base-alpha-3b
 ```
@@ -93,7 +93,7 @@ With only a few modifications, you can prepare and train on your own instruction
 5. Run `finetune_adapter.py` by passing in the location of your data (and optionally other parameters):
    
     ```bash
-    python finetune_adapter.py \
+    python finetune/adapter.py \
         --data_dir data/mydata/ \
         --checkpoint_dir checkpoints/stabilityai/stablelm-base-alpha-3b \
         --out_dir data/mydata-finetuned
