+++ howto/tpus.md
@@ -34,7 +34,7 @@ Since you created a new machine, you'll probably need to download the weights. Y
 Generation works out-of-the-box with TPUs:
 
 ```shell
-python3 generate.py --prompt "Hello, my name is" --num_samples 3
+python3 generate/base.py --prompt "Hello, my name is" --num_samples 3
 ```
 
 This command will take a long time as XLA needs to compile the graph: ~20s for the first generation.
