+++ tests/conftest.py
@@ -6,11 +6,7 @@ import pytest
 wd = Path(__file__).parent.parent.absolute()
 
 
-@pytest.fixture()
-def lit_parrot():
+@pytest.fixture(autouse=True)
+def add_wd_to_path():
     # this adds support for running tests without the package installed
     sys.path.append(str(wd))
-
-    import lit_parrot
-
-    return lit_parrot
