+++ tests/test_model.py
@@ -1,23 +1,11 @@
-import functools
+import sys
 from pathlib import Path
 
-import torch
 import pytest
-import sys
-
+import torch
 from transformers import GPTNeoXForCausalLM, PretrainedConfig
 
-
 wd = Path(__file__).parent.parent.absolute()
-
-
-@functools.lru_cache(maxsize=1)
-def load_convert_script():
-    sys.path.append(str(wd / "scripts"))
-
-    import convert_hf_checkpoint
-
-    return convert_hf_checkpoint
 
 
 @torch.inference_mode()
@@ -26,7 +14,10 @@ def load_convert_script():
 @pytest.mark.parametrize("n_embd", (16, 32))
 @pytest.mark.parametrize("parallel_residual", (False, True))
 @pytest.mark.parametrize("kv_cache", (False, True))
-def test_against_hf_model(rotary_pct, batch_size, n_embd, parallel_residual, kv_cache, lit_parrot) -> None:
+def test_against_hf_model(rotary_pct, batch_size, n_embd, parallel_residual, kv_cache) -> None:
+    import lit_parrot
+    from scripts.convert_hf_checkpoint import copy_weights
+
     block_size = 64
     # https://huggingface.co/stabilityai/stablelm-base-alpha-3b/blob/main/config.json#L24
     vocab_size = 100
@@ -64,9 +55,8 @@ def test_against_hf_model(rotary_pct, batch_size, n_embd, parallel_residual, kv_
     state_dict = ours_model.state_dict()
     theirs_model = GPTNeoXForCausalLM(theirs_config)
 
-    convert_hf_checkpoint = load_convert_script()
     # load the hf initialization into our model
-    convert_hf_checkpoint.copy_weights(state_dict, theirs_model.state_dict())
+    copy_weights(state_dict, theirs_model.state_dict())
     ours_model.load_state_dict(state_dict)
 
     token_sample = torch.randint(0, ours_config.padded_vocab_size, size=(batch_size, block_size), dtype=torch.int64)
@@ -101,7 +91,8 @@ def test_against_hf_model(rotary_pct, batch_size, n_embd, parallel_residual, kv_
 @pytest.mark.skipif(not torch.cuda.is_available(), reason="Requires CUDA")
 @pytest.mark.xfail(raises=AssertionError)  # https://github.com/Lightning-AI/lit-parrot/issues/13
 @torch.inference_mode()
-def test_model_bfloat16(lit_parrot) -> None:
+def test_model_bfloat16() -> None:
+    import lit_parrot
     from lit_parrot.utils import EmptyInitOnDevice
 
     block_size = 64
@@ -131,7 +122,9 @@ def test_model_bfloat16(lit_parrot) -> None:
 
 @pytest.mark.skipif(sys.platform in ("win32", "darwin"), reason="torch.compile not supported on this platform")
 @torch.inference_mode()
-def test_model_compile(lit_parrot):
+def test_model_compile():
+    import lit_parrot
+
     config = lit_parrot.Config(block_size=8, vocab_size=8, n_layer=2, n_head=2, n_embd=4)
     model = lit_parrot.Parrot(config)
     model.apply(model._init_weights)
